+++ tests/testthat/test-demean.R
@@ -175,6 +175,30 @@ test_that("demean for cross-classified designs (by > 1)", {
     tolerance = 1e-4,
     ignore_attr = TRUE
   )
+
+
+  # More than 2 groupings
+  mu <- 100
+  ul <- setNames(c(-1, -3, 0, 4), nm = letters[1:4])
+  uL <- setNames(c(10, 30, 0, -40), nm = LETTERS[1:4])
+  um <- setNames(c(100, 150, -250), nm = month.abb[1:3])
+
+  dat <- expand.grid(l = letters[1:4], L = LETTERS[1:4], m = month.abb[1:3])
+
+  set.seed(111)
+  e <- rnorm(nrow(dat) - 1) |> round(2)
+  e <- append(e, -sum(e))
+
+  dat$y <- mu + ul[dat$l] + uL[dat$L] + um[dat$m] + e
+  dat$z <- mu + ul[dat$l] + uL[dat$L] + um[dat$m] + 10 * e
+
+  dat_dem <- datawizard::demean(dat, by = c("l", "L", "m"), select = c("y", "z"))
+
+  expect_equal(dat_dem$y_l_between, ave(dat$y, dat$l), ignore_attr = TRUE)
+  expect_equal(dat_dem$y_L_between, ave(dat$y, dat$L), ignore_attr = TRUE)
+  expect_equal(dat_dem$y_m_between, ave(dat$y, dat$m), ignore_attr = TRUE)
+  expect_equal(rowSums(dat_dem[grepl("^y_", colnames(dat_dem))]), dat$y)
+  expect_equal(rowSums(dat_dem[grepl("^z_", colnames(dat_dem))]), dat$z)
 })
 
 
@@ -245,4 +269,47 @@ test_that("demean for nested designs (by > 1), nested = TRUE", {
     tolerance = 1e-4,
     ignore_attr = TRUE
   )
+
+  # Following #635
+  testdf <- data.frame(
+    roman = c("I", "I", "I", "I", "II", "II", "II", "II"),
+    alphabet = c("a", "a", "b", "b", "a", "b", "a", "b"),
+    val1 = c(1, 2, 3, 4, 5, 6, 7, 8),
+    val2 = c(1, 2, 3, 4, 5, 6, 7, 8),
+    val3 = c(1, 2, 3, 4, 5, 6, 7, 8)
+  )
+
+  out <- datawizard::demean(testdf,
+    select = c("val1", "val2", "val3"),
+    by = "roman/alphabet", append = FALSE
+  )
+
+  expect_named(out, c(
+    "val1_roman_between", "val1_alphabet_between", "val2_roman_between",
+    "val2_alphabet_between", "val3_roman_between", "val3_alphabet_between",
+    "val1_within", "val2_within", "val3_within"
+  ))
+
+  expect_equal(
+    as.vector(out$val1_within),
+    c(-0.5, 0.5, -0.5, 0.5, -1, -1, 1, 1)
+  )
+  expect_equal(out$val1_within, out$val2_within)
+  expect_equal(out$val1_within, out$val3_within)
+
+  expect_equal(
+    as.vector(out$val1_roman_between),
+    c(2.5, 2.5, 2.5, 2.5, 6.5, 6.5, 6.5, 6.5)
+  )
+  expect_equal(out$val1_roman_between, out$val2_roman_between)
+  expect_equal(out$val1_roman_between, out$val3_roman_between)
+
+  expect_equal(
+    as.vector(out$val1_alphabet_between),
+    c(-1, -1, 1, 1, -0.5, 0.5, -0.5, 0.5)
+  )
+  expect_equal(out$val1_alphabet_between, out$val2_alphabet_between)
+  expect_equal(out$val1_alphabet_between, out$val3_alphabet_between)
+
+  expect_equal(rowSums(out[, grepl("^val1", names(out))]), testdf$val1)
 })