+++ tests/testthat/test-describe_distribution.R
@@ -395,3 +395,12 @@ test_that("(multiple) centralities with CIs", {
   # only one message for data frame
   expect_silent(expect_message(describe_distribution(iris, ci = 0.95)))
 })
+
+
+test_that("display() method exports to markdown", {
+  skip_if_not_installed("knitr")
+  data(iris)
+  out <- describe_distribution(iris)
+  expect_error(display(out, format = "invalid"), regex = "Invalid option")
+  expect_snapshot(display(out))
+})